rename dir `apis` to `api`
