+++ pkg/clientset/versioned/typed/zora/v1alpha1/clusterissue.go
@@ -6,7 +6,7 @@ import (
 	"context"
 	"time"
 
-	v1alpha1 "github.com/undistro/zora/apis/zora/v1alpha1"
+	v1alpha1 "github.com/undistro/zora/api/zora/v1alpha1"
 	scheme "github.com/undistro/zora/pkg/clientset/versioned/scheme"
 	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	types "k8s.io/apimachinery/pkg/types"
