+++ pkg/clientset/versioned/typed/zora/v1alpha1/fake/fake_clusterissue.go
@@ -5,7 +5,7 @@ package fake
 import (
 	"context"
 
-	v1alpha1 "github.com/undistro/zora/apis/zora/v1alpha1"
+	v1alpha1 "github.com/undistro/zora/api/zora/v1alpha1"
 	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	labels "k8s.io/apimachinery/pkg/labels"
 	schema "k8s.io/apimachinery/pkg/runtime/schema"
