+++ pkg/clientset/versioned/typed/zora/v1alpha1/zora_client.go
@@ -5,7 +5,7 @@ package v1alpha1
 import (
 	"net/http"
 
-	v1alpha1 "github.com/undistro/zora/apis/zora/v1alpha1"
+	v1alpha1 "github.com/undistro/zora/api/zora/v1alpha1"
 	"github.com/undistro/zora/pkg/clientset/versioned/scheme"
 	rest "k8s.io/client-go/rest"
 )
