+++ pkg/worker/report/parse.go
@@ -21,7 +21,7 @@ import (
 
 	"github.com/go-logr/logr"
 
-	zorav1a1 "github.com/undistro/zora/apis/zora/v1alpha1"
+	zorav1a1 "github.com/undistro/zora/api/zora/v1alpha1"
 	"github.com/undistro/zora/pkg/worker/config"
 
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
