+++ pkg/worker/run/run.go
@@ -25,7 +25,7 @@ import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	ctrl "sigs.k8s.io/controller-runtime"
 
-	zorav1a1 "github.com/undistro/zora/apis/zora/v1alpha1"
+	zorav1a1 "github.com/undistro/zora/api/zora/v1alpha1"
 	"github.com/undistro/zora/pkg/clientset/versioned"
 	"github.com/undistro/zora/pkg/worker/config"
 	"github.com/undistro/zora/pkg/worker/report"
